last to master by phonphey

+++ main.py
@@ -6,18 +6,18 @@ def main():
     type = sys.argv[1]
     
     if type == "planets":
-        gen.gen_planets(int(sys.argv[2]), int(sys.argv[3]))
+        gen.gen_planets(int(sys.argv[2]), float(sys.argv[3]))
 
     elif type == "comets":
-        gen.gen_comets(int(sys.argv[2]), int(sys.argv[3]))
+        gen.gen_comets(int(sys.argv[2]), float(sys.argv[3]))
     
     elif type == "stars":
-        gen.gen_stars(int(sys.argv[2]), int(sys.argv[3]))
+        gen.gen_stars(int(sys.argv[2]), float(sys.argv[3]))
     
     elif type == "all":
-        gen.gen_stars(int(sys.argv[2]), int(sys.argv[3]))
-        gen.gen_comets(int(sys.argv[2]), int(sys.argv[3]))
-        gen.gen_planets(int(sys.argv[2]), int(sys.argv[3]))
+        gen.gen_stars(int(sys.argv[2]), float(sys.argv[3]))
+        gen.gen_comets(int(sys.argv[2]), float(sys.argv[3]))
+        gen.gen_planets(int(sys.argv[2]), float(sys.argv[3]))
     
     else:
         print("%s type is not defined" % type)
